requirements.txt
# Core dependencies for website cloning from screenshots
selenium==4.15.2
Pillow==10.1.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
lxml==4.9.3

# Optional: For AI-powered conversion
openai==1.3.5
anthropic==0.7.7
transformers==4.35.2

# Development tools
python-dotenv==1.0.0
